PriceMapProject/backend/requirements.txt
asgiref>=3.8.1
Django==5.2
django-cors-headers==4.3.1
djangorestframework==3.15.0
pillow==10.2.0
python-dotenv==1.0.0
sqlparse==0.4.4
tzdata==2024.1
